fix(mobile): play the brand splash on every login, not just the next cold start (#500)
The animated brand splash was an overlay in the root layout, gated on an
AsyncStorage `pending-splash` flag read once in a mount-time useEffect.
login/signup set the flag then `router.replace` into the app — but in-session
navigation doesn't remount the root layout, so the flag was never re-read and
the splash only played on the NEXT cold launch, never right after login.

Move it to its own route, app/(auth)/welcome.tsx: login/signup `router.replace`
there (dropping the flag), it plays the staged fade, then forwards into /(app).
Plays after every login by construction.

- new app/(auth)/welcome.tsx — staged-fade brand splash + tap-to-skip +
  reduced-motion, then router.replace('/(app)').
- _layout.tsx — stripped the overlay/flag/animation; now only holds the native
  splash for fonts + a loading gate over the cold-start auth-resolve window.
- login.tsx / signup.tsx — replace('/(auth)/welcome'); removed the now-unused
  AsyncStorage flag.
